calamari_common: Move Event into common
