+++ calamari-common/setup.py
@@ -3,7 +3,7 @@ from setuptools import setup
 setup(
     name="calamari_common",
     version="0.1",
-    packages=['calamari_common'],
+    packages=['calamari_common', 'calamari_common.db'],
     url="http://www.inktank.com/enterprise/",
     author="Inktank Storage Inc.",
     author_email="info@inktank.com",
